refactor(post-create): extract filters type & array
